solution_micaela/requirements.txt
langchain==0.0.300
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
openai==1.0.0
tqdm==4.65.0
pandas==2.2.3
scikit-learn==1.6.1
joblib==1.3.2
